Modifico el script de teardown para avisar que no hay containers de Andino corriendo

+++ scripts/tear_down_services.sh
@@ -4,5 +4,10 @@ if [[ -f /etc/portal/latest.yml ]];
 then
     cd /etc/portal && docker-compose -f /etc/portal/latest.yml down && cd .. && sudo rm -rf /etc/portal/
 else
-    docker kill $(docker ps -q -f name=andino)
+    if [[ $(docker ps -q -f name=andino) ]];
+    then
+        docker kill $(docker ps -q -f name=andino)
+    else
+        echo "No hay contenedores de Andino corriendo."
+    fi
 fi
